+++ packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/AuthorAndPublishInfo/AuthorAndPublishInfo.scss
@@ -63,7 +63,7 @@
     .CWSelectList {
       min-width: 202px;
       margin-left: -20px;
-      margin-right: -10px;
+      margin-right: -8px;
 
       .SelectList:not(:focus):not(:focus-within) {
         border-color: transparent !important;
